Add fetchTodoData action
